+++ src/lib/ultimem/ultimem.h
@@ -25,6 +25,6 @@ extern void __fastcall__ ultimem_copy_rom2ram (long src, long dst, unsigned size
 
 // Map pointer into external bank configuration to another block
 // in the current configuration.
-extern void * __fastcall__ ultimem_map_ptr (void * block, int * blockreg, void * ptr, int * blockregs);
+extern void * __fastcall__ ultimem_map_ptr (void * block, unsigned short * blockreg, void * ptr, unsigned short * blockregs);
 
 #endif /* #define ULTIMEM_H */
